requirements/dev.txt
-r base.txt
django-debug-toolbar==1.2.1
django-extensions==1.4.6
pip-tools==0.3.5
six==1.8.0
sqlparse==0.1.13
Werkzeug==0.9.6
